health-mcp/requirements.txt
mcp[cli]
jsonschema
python-dateutil